+++ phonemizer/main.py
@@ -213,14 +213,15 @@ Exemples:
         "sampa" (Speech Assessment Methods Phonetic Alphabet) alphabet instead
         of "ipa" (International Phonetic Alphabet).''')
     group.add_argument(
-        '--language-switch', default='keep-flags',
-        choices=['keep-flags', 'remove-flags', 'remove-utterance'],
+        '--language-switch', default='ignore',
+        choices=['remove-utterance', 'ignore', 'remove-flags', 'keep-flags'],
         help="""espeak can pronounce some words in another language (typically
         English) when phonemizing a text. This option setups the policy to use
-        when such a language switch occurs. Three values are available:
-        'keep-flags' (the default), 'remove-flags' or 'remove-utterance'. The
-        'keep-flags' policy keeps the language switching flags, for example
-        (en) or (jp), in the output. The 'remove-flags' policy removes them and
+        when such a language switch occurs. Four values are available:
+        'ignore' (the default), 'keep-flags','remove-flags' or 'remove-utterance'. 
+        The 'ignore' policy disable language switching,'keep-flags' policy keeps 
+        the language switching flags, for example (en) or (jp), in the output. 
+        The 'remove-flags' policy removes the flags (en) (jp), and 
         the 'remove-utterance' policy removes the whole line of text including
         a language switch.""")